server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "blinker",
    "click",
    "flask",
    "flask-sock",
    "gevent",
    "greenlet",
    "h11",
    "itsdangerous",
    "jinja2",
    "markupsafe",
    "pytest>=9.0.1",
    "python-dotenv",
    "python-magic>=0.4.27",
    "setuptools",
    "simple-websocket",
    "websockets",
    "werkzeug",
    "wsproto",
    "zope-event",
    "zope-interface",
    "redis>=5.0.7",
    "bcrypt>=4.0.0",
    "flask-cors>=4.0.0",
    "marshmallow>=3.20.0",
    "coverage[toml]>=7.4.0",
]

[tool.coverage.run]
branch = true
source = ["server"]
omit = ["server/tests/*"]

[tool.coverage.report]
show_missing = true
skip_covered = true
